md2pptx/requirements.txt
python-pptx>=1.0.0
matplotlib>=3.5.0
Pillow>=9.0.0
markdown-it-py>=3.0.0
google-genai>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
